Değişiklikler:
- file ve img utils dosyaları oluşturuldu
- output klasörü silindi
- save_img fonksiyonuna otomatik klasör oluşturma kodu yazıldı
