+++ detect.py
@@ -1,6 +1,7 @@
-from utils.datasets import *
+from utils.datasets import numpy_to_torch, torch_to_numpy, coco_classes
 from utils.object_counter import counter
 from utils.vis import object_visulization
+from utils.img_utils import read_image, save_image, resize_image
 import cv2
 import torchvision
 import torch
