+++ doc/bird_out.png
