+++ doc/mix_out.png
