+++ utils/datasets.py
@@ -1,5 +1,3 @@
-import cv2
-import numpy as np
 import torch
 
 coco_classes = (
@@ -18,29 +16,6 @@ coco_classes = (
 )
 
 
-def read_image(img):
-    if type(img) == str:
-        img = cv2.imread(img)
-        # img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-
-    elif type(img) == bytes:
-        nparr = np.frombuffer(img, np.uint8)
-        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-
-    elif type(img) == np.ndarray:
-        if len(img.shape) == 2:  # grayscale
-            img = cv2.cvtColor(img, cv2.COLOR_GRAY2RGB)
-
-        elif len(img.shape) == 3 and img.shape[2] == 3:
-            img = img
-
-        elif len(img.shape) == 3 and img.shape[2] == 4:  # RGBA
-            img = img[:, :, :3]
-
-    return img
-
-
 def numpy_to_torch(img):
     img = img.transpose((2, 0, 1))
     img = torch.from_numpy(img).float()
@@ -55,22 +30,3 @@ def torch_to_numpy(img):
         img /= 255
     img = img.transpose((1, 2, 0))
     return img
-
-
-def save_image(img, path):
-    img = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
-    cv2.imwrite(path, img)
-
-
-def resize_image(image, long_size, interpolation=cv2.INTER_LINEAR):
-    height, width, channel = image.shape
-
-    # set target image size
-    target_size = long_size
-
-    ratio = target_size / max(height, width)
-
-    target_h, target_w = int(height * ratio), int(width * ratio)
-    image = cv2.resize(image, (target_w, target_h), interpolation=interpolation)
-
-    return image
